Added health to construtor and turned health to float

+++ core/src/com/mygdx/gigabiteconomy/sprites/IHealthBar.java
@@ -4,5 +4,7 @@ import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 
 public interface IHealthBar {
     void drawOn(SpriteBatch batch);
-    void modifyHealth(int dhealth);
+
+    void modifyHealth(float dhealth);
+
 }
